fixed partial registration issue
