+++ app/services/bike_creator_verifier.rb
@@ -23,7 +23,7 @@ class BikeCreatorVerifier
     @bike.phone ||= @b_param.params['stolen_record']['phone'] if @b_param.params && @b_param.params['stolen_record'].present?
     if @bike.creation_organization.present? && @bike.creation_organization.locations.any?
       @bike.phone ||= @bike.creation_organization.locations.first.phone
-    elsif @bike.creator.phone.present?
+    elsif @bike.creator && @bike.creator.phone.present?
       @bike.phone ||= @bike.creator.phone
     end
   end
